final version

+++ srcs/philosopher_action.c
@@ -6,7 +6,7 @@
 /*   By: nwyseur <nwyseur@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/31 15:43:21 by nwyseur           #+#    #+#             */
-/*   Updated: 2023/06/01 12:19:51 by nwyseur          ###   ########.fr       */
+/*   Updated: 2023/06/13 14:38:04 by nwyseur          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ int	ft_forks(t_philo *philo, pthread_mutex_t *r_fork, pthread_mutex_t *l_fork)
 	if (ft_watch_death(philo) == 1)
 		return (pthread_mutex_unlock(r_fork), 0);
 	pthread_mutex_lock(&philo->table->printf);
-	printf("\n%i %d has taken a fork",
+	printf("\n%lu %d has taken a fork",
 		ft_get_time() - philo->table->time_start, philo->id);
 	pthread_mutex_unlock(&philo->table->printf);
 	pthread_mutex_lock(l_fork);
@@ -26,7 +26,7 @@ int	ft_forks(t_philo *philo, pthread_mutex_t *r_fork, pthread_mutex_t *l_fork)
 		return (pthread_mutex_unlock(r_fork),
 			pthread_mutex_unlock(l_fork), 0);
 	pthread_mutex_lock(&philo->table->printf);
-	printf("\n%i %d has taken a fork",
+	printf("\n%lu %d has taken a fork",
 		ft_get_time() - philo->table->time_start, philo->id);
 	pthread_mutex_unlock(&philo->table->printf);
 	return (1);
@@ -56,7 +56,7 @@ int	ft_eat(t_philo *philo)
 	pthread_mutex_lock(&philo->table->printf);
 	philo->eat_count++;
 	philo->last_eat_time = ft_get_time();
-	printf("\n%i %d is eating",
+	printf("\n%lu %d is eating",
 		ft_get_time() - philo->table->time_start, philo->id);
 	pthread_mutex_unlock(&philo->table->printf);
 	ft_usleep_check_death(philo, philo->table->time_eat);
@@ -78,7 +78,7 @@ int	ft_sleep(t_philo *philo)
 	if (ft_watch_death(philo) == 1)
 		return (0);
 	pthread_mutex_lock(&philo->table->printf);
-	printf("\n%i %d is sleeping",
+	printf("\n%lu %d is sleeping",
 		ft_get_time() - philo->table->time_start, philo->id);
 	pthread_mutex_unlock(&philo->table->printf);
 	ft_usleep_check_death(philo, philo->table->time_sleep);
@@ -87,10 +87,10 @@ int	ft_sleep(t_philo *philo)
 
 int	ft_think(t_philo *philo)
 {
-	int	time;
-	int	cycle;
-	int	nb_philo;
-	int	die;
+	unsigned long	time;
+	unsigned long	cycle;
+	int				nb_philo;
+	unsigned long	die;
 
 	nb_philo = philo->table->nb_philo;
 	die = philo->table->time_die;
@@ -105,7 +105,7 @@ int	ft_think(t_philo *philo)
 	if (ft_watch_death(philo) == 1)
 		return (0);
 	pthread_mutex_lock(&philo->table->printf);
-	printf("\n%i %d is thinking",
+	printf("\n%lu %d is thinking",
 		ft_get_time() - philo->table->time_start, philo->id);
 	pthread_mutex_unlock(&philo->table->printf);
 	ft_usleep_check_death(philo, time);
